Add middlewares, change to controllers
