+++ controllers/contacts/getAllContacts.js
@@ -1,0 +1,8 @@
+const contacts = require("../../models/contacts/index");
+
+const getAllContacts = async (req, res) => {
+  const result = await contacts.listContacts();
+  res.json(result);
+};
+
+module.exports = getAllContacts;
