+++ controllers/contacts/getContactById.js
@@ -1,0 +1,14 @@
+const { NotFound } = require("http-errors");
+
+const contacts = require("../../models/contacts");
+
+const getContactById = async (req, res) => {
+  const { contactId } = req.params;
+  const result = await contacts.getContactById(contactId);
+  if (!result) {
+    throw new NotFound();
+  }
+  res.json(result);
+};
+
+module.exports = getContactById;
