+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const getAllContacts = require("./getAllContacts");
+const getContactById = require("./getContactById");
+const postContact = require("./postContact");
+const deleteContact = require("./deleteContact");
+const putContact = require("./putContact");
+
+module.exports = {
+  getAllContacts,
+  getContactById,
+  postContact,
+  deleteContact,
+  putContact,
+};
