+++ db/contacts.json
@@ -69,13 +69,13 @@
     "id": "1e422f6d-d35a-40e6-aeb7-cc876bfbeb7a",
     "name": "Simon Morton",
     "email": "dui.Fusce.diam@Donec.com",
-    "phone": "(233) 738-9990"
+    "phone": "(233) 738-9991"
   },
   {
     "id": "8c389657-5d80-40c1-bc75-818ddb3ec938",
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
-    "phone": "(748) 206-2688"
+    "phone": "(748) 206-2558"
   },
   {
     "id": "791b3ad6-0b6c-40c4-87c6-74fdcd21d3d7",
