+++ middlewares/index.js
@@ -1,0 +1,7 @@
+const validation = require("./validation");
+const ctrlWrapper = require("./ctrlWrapper");
+
+module.exports = {
+  validation,
+  ctrlWrapper,
+};
