+++ models/contacts/contactsPath.js
@@ -1,0 +1,5 @@
+const path = require("path");
+
+const contactsPath = path.join("db", "contacts.json");
+
+module.exports = contactsPath;
