+++ models/contacts/removeContact.js
@@ -1,0 +1,18 @@
+const fs = require("fs/promises");
+
+const listContacts = require("./listContacts");
+const contactsPath = require("./contactsPath");
+
+const removeContact = async (contactId) => {
+  const contacts = await listContacts();
+  const idx = contacts.findIndex((item) => item.id === contactId);
+  if (idx === -1) {
+    return null;
+  }
+  const deleteContact = contacts[idx];
+  contacts.splice(idx, 1);
+  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
+  return deleteContact;
+};
+
+module.exports = removeContact;
