+++ models/contacts/updateContact.js
@@ -1,0 +1,18 @@
+const fs = require("fs/promises");
+
+const listContacts = require("./listContacts");
+const contactsPath = require("./contactsPath");
+
+const updateContact = async (contactId, body) => {
+  const contacts = await listContacts();
+  const idx = contacts.findIndex((item) => item.id === contactId);
+  if (idx === -1) {
+    return null;
+  }
+
+  contacts[idx] = { id: contactId, ...body };
+  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
+  return contacts[idx];
+};
+
+module.exports = updateContact;
