+++ routes/api/contacts.js
@@ -1,83 +1,21 @@
 const express = require("express");
-const { BadRequest, NotFound } = require("http-errors");
-const Joi = require("joi");
-
-const contacts = require("../../models/contacts");
-
-const contactsSchema = Joi.object({
-  name: Joi.string().min(3).max(20).required(),
-  email: Joi.string()
-    .email({ minDomainSegments: 2, tlds: { allow: ["com", "net", "ua"] } })
-    .required(),
-  phone: Joi.string().max(14).required(),
-});
-
 const router = express.Router();
 
-router.get("/", async (req, res, next) => {
-  try {
-    const result = await contacts.listContacts();
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+const { validation, ctrlWrapper } = require("../../middlewares");
+const { contactSchema } = require("../../schemas");
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.getContactById(contactId);
-    if (!result) {
-      throw new NotFound();
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+const ctrl = require("../../controllers/contacts");
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = contactsSchema.validate(req.body);
-    if (error) {
-      throw new BadRequest("missing required field");
-    }
-    const result = await contacts.addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+const validateMiddleware = validation(contactSchema);
 
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.removeContact(contactId);
+router.get("/", ctrlWrapper(ctrl.getAllContacts));
 
-    if (!result) {
-      throw new NotFound();
-    }
-    res.json({ message: "Contact deleted" });
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:contactId", ctrlWrapper(ctrl.getContactById));
 
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const { error } = contactsSchema.validate(req.body);
-    if (error) {
-      throw new BadRequest("missing fields");
-    }
-    const { contactId } = req.params;
-    const result = await contacts.updateContact(contactId, req.body);
-    if (!result) {
-      throw new NotFound();
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", validateMiddleware, ctrlWrapper(ctrl.postContact));
+
+router.delete("/:contactId", ctrlWrapper(ctrl.deleteContact));
+
+router.put("/:contactId", validateMiddleware, ctrlWrapper(ctrl.putContact));
 
 module.exports = router;
