+++ schemas/contact.js
@@ -1,0 +1,11 @@
+const Joi = require("joi");
+
+const contactSchema = Joi.object({
+  name: Joi.string().min(3).max(20).required(),
+  email: Joi.string()
+    .email({ minDomainSegments: 2, tlds: { allow: ["com", "net", "ua"] } })
+    .required(),
+  phone: Joi.string().max(14).required(),
+});
+
+module.exports = contactSchema;
